removed unnecessary casts
git-svn-id: http://svn.vuze.com/public/client@142 c2c1b7af-1477-405b-86ee-8aefd1fc4dd2

+++ trunk/azureus2/src/org/gudy/azureus2/core/MessageDigest.java
@@ -312,7 +312,7 @@ public abstract class MessageDigest {
                 md.update(testdata);
             md.finish();
             times[i] = System.currentTimeMillis() - t1;
-            rates[i] = ((double) (i*testdata.length) * 1000.0)/ (double) times[i];
+            rates[i] = ((i*testdata.length) * 1000.0) / times[i];
             System.out.println((i*testdata.length)+"\t"+times[i]+"\t"+rates[i]);
         }
         System.out.println("Done.");
